Use hoook module path
